Add files via upload
Breast cancer is a cancer that forms in the cells of the breasts.
It can occur in women and rarely in men.
Symptoms of breast cancer include a lump in the breast, bloody discharge from the nipple and changes in the shape or texture of the nipple or breast. Treatment depends on the stage of cancer. It may consist of chemotherapy, radiation and surgery.

It's important to understand that most breast lumps are benign and not cancer (malignant). Non-cancerous breast tumors are abnormal growths, but they do not spread outside of the breast. They are not life threatening, but some types of benign breast lumps can increase a woman's risk of getting breast cancer. Which is why it is important for us to create a predictive model to ease the researchers works.

This Predictive Analytics was carried out with 'R'